Update install-dashy-alpine3.16.sh
Added line to create user "dashy" 
Added line to create user group "dashy"
Added args to boot script to assign it to user "dashy"

This is an attempt to fix a problem where yarn is not "auto building" changes made to the config on the fly. It seems this may be an issue when running Yarn as root or a root user.